[ossvm] OSSVM implementation for binary classification
This commit provides the OSSVM implementation as proposed by [1].  For
usage through LibSVM’s executable, it provides the additional `-l`
option for the lambda regularization parameters described in
Equation (11) of the paper [1].  The `-l` option is only accepted when
the kernel type is RBF, i.e., when `-t 2` is provided.

Note: Although in the paper [1] it is presented that the lambda
regularization parameter can be up to `C * mp`, in which `C` is the
SVM cost parameter and `mp` is the number of positive training
instances on the training set, to facilitate usage, `-l` option
accepts a value in the interval `[0, 1)` which is linearly mapped to
the interval `[0, C*mp)`.  Value of the lambda parameters equals to 1
is not accepted, as explained in the second paragraph after
Proposition 1 in the paper [1].

This commit also adjusts the Python wrap for usage.

[1] Mendes Júnior, Pedro Ribeiro; Boult, Terrance E.; Wainer, Jacques;
and Rocha, Anderson de Rezende (2021), “Open-Set Support Vector
Machines”, to appear in IEEE Transactions on Systems, Man, and
Cybernetics: Systems. URL: https://pedrormjunior.github.io/OSSVM.html